backend/requirements-cv.txt
numpy
Pillow
scikit-learn
tensorflow
kagglehub
